[FEATURE] Implemented a PeptideIndexing class with a function equivalent to the PeptideIndexer tool
