+++ src/openms/include/OpenMS/ANALYSIS/ID/PeptideIndexing.h
@@ -1,0 +1,89 @@
+// --------------------------------------------------------------------------
+//                   OpenMS -- Open-Source Mass Spectrometry
+// --------------------------------------------------------------------------
+// Copyright The OpenMS Team -- Eberhard Karls University Tuebingen,
+// ETH Zurich, and Freie Universitaet Berlin 2002-2015.
+//
+// This software is released under a three-clause BSD license:
+//  * Redistributions of source code must retain the above copyright
+//    notice, this list of conditions and the following disclaimer.
+//  * Redistributions in binary form must reproduce the above copyright
+//    notice, this list of conditions and the following disclaimer in the
+//    documentation and/or other materials provided with the distribution.
+//  * Neither the name of any author or any participating institution
+//    may be used to endorse or promote products derived from this software
+//    without specific prior written permission.
+// For a full list of authors, refer to the file AUTHORS.
+// --------------------------------------------------------------------------
+// THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
+// AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
+// IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
+// ARE DISCLAIMED. IN NO EVENT SHALL ANY OF THE AUTHORS OR THE CONTRIBUTING
+// INSTITUTIONS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL,
+// EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO,
+// PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS;
+// OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY,
+// WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR
+// OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF
+// ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+//
+// --------------------------------------------------------------------------
+// $Maintainer: Chris Bielow $
+// $Authors: Andreas Bertsch, Chris Bielow, Knut Reinert $
+// --------------------------------------------------------------------------
+
+#include <OpenMS/DATASTRUCTURES/DefaultParamHandler.h>
+#include <OpenMS/KERNEL/StandardTypes.h>
+#include <OpenMS/CONCEPT/ProgressLogger.h>
+#include <OpenMS/FORMAT/FASTAFile.h>
+
+namespace OpenMS
+{
+
+ class OPENMS_DLLAPI PeptideIndexing :
+    public DefaultParamHandler,
+    public ProgressLogger
+  {
+public:
+
+    /// Exit codes
+    enum ExitCodes
+    {
+      EXECUTION_OK,
+      DATABASE_EMPTY,
+      PEPTIDE_IDS_EMPTY,
+      DATABASE_CONTAINS_MULTIPLES,
+      ILLEGAL_PARAMETERS,
+      UNEXPECTED_RESULT
+    };
+
+    /// Default constructor
+    PeptideIndexing();
+
+    /// Default destructor
+    virtual ~PeptideIndexing();
+
+    /// main method of FeatureFindingMetabo
+    ExitCodes run(std::vector<FASTAFile::FASTAEntry>& proteins, std::vector<ProteinIdentification>& prot_ids, std::vector<PeptideIdentification>& pep_ids);
+
+protected:
+    virtual void updateMembers_();
+
+    String decoy_string_;
+    String prefix_;
+    String missing_decoy_action_;
+    String enzyme_name_;
+    String enzyme_specificity_;
+
+    bool write_protein_sequence_;
+    bool write_protein_description_;
+    bool keep_unreferenced_proteins_;
+    bool allow_unmatched_;
+    bool full_tolerant_search_;
+    bool IL_equivalent_;
+
+    Size aaa_max_;
+    Size mismatches_max_;
+
+  };
+}
