+++ src/openms/include/OpenMS/ANALYSIS/ID/sources.cmake
@@ -28,6 +28,7 @@ PILISNeutralLossModel.h
 PILISScoring.h
 PILISCrossValidation.h
 ProtonDistributionModel.h
+PeptideIndexing.h
 )
 
 ### add path to the filenames
